techpack: display: dsi: Force AOD brightness to low
The AOD_HIGH is too bright in a dark room.

Sony has implements an auto brightess AOD feature in its framework,
and in stock it will adjust the brightness value when in doze mode.
However, we don't have this feature and the default HIGH brightness is
too bright so we adjust the threshold to force the screen to enter
AOD_LOW mode.

Change-Id: Ie755fb2ebc5653f3577fffa4769bd9561ffa8eda
Signed-off-by: SirRGB <sirrgb@proton.me>